Project 05
# project 5
The purpose of this project is to create a 3D game with a first-person perspective. To score 15 points, you will need to have:

A first-person player-controlled character that can move with keyboard and mouse control
A maze for your character to navigate
At least one object to pick up or interact with
At least one enemy to impede your progress
A use of 2 materials
A use of 2 light sources
A main menu, and an end-game screen
Your game must have a clear objective

## Implementation
Different (especially student-created) models, I created a key model.
Music and sound effects horror sounds effect at start and in-game
Extra enemies add 2 more enemies
Creative use of lights or materials, I made light follow the key, and will flash when player apporch to key
Attention paid to the start- and end-game experience, I made different sound effect for this